# docs/api/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "Hotel competitor monitor API",
  "version": "0.1.0"
 },
 "paths": {
...
  "/market/reputation": {
   "get": {
    "tags": [
     "reputation"
    ],
    "summary": "Reputation",
    "operationId": "reputation_market_reputation_get",
    "parameters": [
     {
      "name": "days",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "maximum": 730,
       "minimum": 7,
       "default": 90,
       "title": "Days"
      }
     },
     {
      "name": "tenant_id",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "integer"
        },
        {
         "type": "null"
        }
       ],
       "description": "Operator: tenant cần xem",
       "title": "Tenant Id"
      },
      "description": "Operator: tenant cần xem"
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ReputationOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/market/visibility": {
   "get": {
    "tags": [
     "reputation"
    ],
    "summary": "Visibility",
    "description": "Vị trí trên trang kết quả của khu vực (thứ tự mặc định của kênh) cho bạn và đối thủ.",
    "operationId": "visibility_market_visibility_get",
    "parameters": [
     {
      "name": "days",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "maximum": 120,
       "minimum": 1,
       "default": 14,
       "title": "Days"
      }
     },
     {
      "name": "area_id",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "integer"
        },
        {
         "type": "null"
        }
       ],
       "title": "Area Id"
      }
     },
     {
      "name": "tenant_id",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "integer"
        },
        {
         "type": "null"
        }
       ],
       "description": "Operator: tenant cần xem",
       "title": "Tenant Id"
      },
      "description": "Operator: tenant cần xem"
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/VisibilityOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
...
   "ReputationHotelOut": {
    "properties": {
     "hotel_id": {
      "type": "integer",
      "title": "Hotel Id"
     },
     "name": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Name"
     },
     "role": {
      "type": "string",
      "title": "Role"
     },
     "review_score": {
      "anyOf": [
       {
        "type": "string",
        "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
       },
       {
        "type": "null"
       }
      ],
      "title": "Review Score"
     },
     "review_count": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Review Count"
     },
     "reviews_per_month": {
      "anyOf": [
       {
        "type": "string",
        "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
       },
       {
        "type": "null"
       }
      ],
      "title": "Reviews Per Month"
     },
     "score_change": {
      "anyOf": [
       {
        "type": "string",
        "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
       },
       {
        "type": "null"
       }
      ],
      "title": "Score Change"
     },
     "next_threshold": {
      "anyOf": [
       {
        "type": "string",
        "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
       },
       {
        "type": "null"
       }
      ],
      "title": "Next Threshold"
     },
     "gap_to_next": {
      "anyOf": [
       {
        "type": "string",
        "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
       },
       {
        "type": "null"
       }
      ],
      "title": "Gap To Next"
     },
     "badges": {
      "items": {
       "type": "string"
      },
      "type": "array",
      "title": "Badges"
     },
     "price_index": {
      "anyOf": [
       {
        "type": "string",
        "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
       },
       {
        "type": "null"
       }
      ],
      "title": "Price Index"
     },
     "score_index": {
      "anyOf": [
       {
        "type": "string",
        "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
       },
       {
        "type": "null"
       }
      ],
      "title": "Score Index"
     }
    },
    "type": "object",
    "required": [
     "hotel_id",
     "name",
     "role",
     "review_score",
     "review_count",
     "reviews_per_month",
     "score_change",
     "next_threshold",
     "gap_to_next",
     "badges",
     "price_index",
     "score_index"
    ],
    "title": "ReputationHotelOut"
   },
   "ReputationOut": {
    "properties": {
     "days": {
      "type": "integer",
      "title": "Days"
     },
     "hotels": {
      "items": {
       "$ref": "#/components/schemas/ReputationHotelOut"
      },
      "type": "array",
      "title": "Hotels"
     }
    },
    "type": "object",
    "required": [
     "days",
     "hotels"
    ],
    "title": "ReputationOut"
   },
...
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "type": "array",
      "title": "Location"
     },
     "msg": {
      "type": "string",
      "title": "Message"
     },
     "type": {
      "type": "string",
      "title": "Error Type"
     },
     "input": {
      "title": "Input"
     },
     "ctx": {
      "type": "object",
      "title": "Context"
     }
    },
    "type": "object",
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError"
   },
   "VisibilityHotelOut": {
    "properties": {
     "hotel_id": {
      "type": "integer",
      "title": "Hotel Id"
     },
     "name": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Name"
     },
     "role": {
      "type": "string",
      "title": "Role"
     },
     "scans": {
      "type": "integer",
      "title": "Scans"
     },
     "avg_rank": {
      "anyOf": [
       {
        "type": "string",
        "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
       },
       {
        "type": "null"
       }
      ],
      "title": "Avg Rank"
     },
     "avg_organic_rank": {
      "anyOf": [
       {
        "type": "string",
        "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
       },
       {
        "type": "null"
       }
      ],
      "title": "Avg Organic Rank"
     },
     "best_rank": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Best Rank"
     },
     "sponsored_share": {
      "anyOf": [
       {
        "type": "string",
        "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
       },
       {
        "type": "null"
       }
      ],
      "title": "Sponsored Share"
     },
     "badges": {
      "items": {
       "type": "string"
      },
      "type": "array",
      "title": "Badges"
     }
    },
    "type": "object",
    "required": [
     "hotel_id",
     "name",
     "role",
     "scans",
     "avg_rank",
     "avg_organic_rank",
     "best_rank",
     "sponsored_share",
     "badges"
    ],
    "title": "VisibilityHotelOut"
   },
   "VisibilityOut": {
    "properties": {
     "area_id": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Area Id"
     },
     "days": {
      "type": "integer",
      "title": "Days"
     },
     "hotels": {
      "items": {
       "$ref": "#/components/schemas/VisibilityHotelOut"
      },
      "type": "array",
      "title": "Hotels"
     }
    },
    "type": "object",
    "required": [
     "area_id",
     "days",
     "hotels"
    ],
    "title": "VisibilityOut"
   },
...
  }
 }
}